requirements.txt
modal
fastapi
opencv-python
uvicorn[standard]
websockets
vllm
transformers
torch
torchvision
ultralytics
chromadb
Pillow
numpy
scipy
requests
python-dotenv
sounddevice
silero-vad
faster-whisper
fpdf2
